fix(test): standardize fresh fixtures across HTTP tests
